requirements/requirements.txt
django==4.0.2
psycopg2==2.9.3 # windows use the downloaded wheel file
djangorestframework==3.13.1
django-cors-headers==3.11.0
django-extensions==3.1.5
PyJWT==2.3.0
django-guardian==2.4.0
django-phonenumber-field==6.1.0
python-dotenv==0.19.2
phonenumberslite==8.12.45
lxml==4.8.0
dateparser==1.1.0
celery==5.2.3
python-redis-lock==3.7.0